pyproject.toml
[project]
name = "ai-resume-critique"
version = "0.1.0"
description = "AI-powered resume analysis tool using Groq"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "requests>=2.32.0",
]